Connectinno-App/backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.1
python-dotenv==1.0.1

firebase-admin==6.5.0
google-cloud-firestore==2.17.0
httpx==0.27.2
